+++ thirdparty/vars.sh
@@ -453,10 +453,10 @@ FAST_FLOAT_SOURCE=fast_float-3.9.0
 FAST_FLOAT_MD5SUM="5656b0d8b150a3b157cfb092d214f6ea"
 
 # libhdfs
-HADOOP_LIBS_X86_DOWNLOAD="https://github.com/apache/doris-thirdparty/releases/download/hadoop-3.3.4.1-for-doris/hadoop_lib_3.3.4.1-x86.tar.gz"
-HADOOP_LIBS_X86_NAME="hadoop_lib_3.3.4.1-x86.tar.gz"
-HADOOP_LIBS_X86_SOURCE="hadoop_lib_3.3.4.1-x86"
-HADOOP_LIBS_X86_MD5SUM="9b434ae90f17a358c55b5476841cb836"
+HADOOP_LIBS_DOWNLOAD="https://github.com/apache/doris-thirdparty/archive/refs/tags/hadoop-3.3.4.1-for-doris.tar.gz"
+HADOOP_LIBS_NAME="hadoop-3.3.4.1-for-doris.tar.gz"
+HADOOP_LIBS_SOURCE="hadoop-3.3.4.1-for-doris"
+HADOOP_LIBS_MD5SUM="d10881a7cf42d94f76db3fb0e3c6e0c0"
 
 # all thirdparties which need to be downloaded is set in array TP_ARCHIVES
 export TP_ARCHIVES=(
@@ -524,7 +524,7 @@ export TP_ARCHIVES=(
     'XXHASH'
     'CONCURRENTQUEUE'
     'FAST_FLOAT'
-    'HADOOP_LIBS_X86'
+    'HADOOP_LIBS'
 )
 
 if [[ "$(uname -s)" == 'Darwin' ]]; then